+++ lib/bitstampede/entities/Order.rb
